refactor Leader controller
